+++ sample.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func main() {
+	x := 5
+	fmt.Printf("x %#v\n", x)
+}