+++ sample.jsx
@@ -1,0 +1,3 @@
+const Foo = ({ name }) => {
+  return <div> hello, {name}</div>;
+};