+++ sample.rb
@@ -1,0 +1,2 @@
+foo = 5
+pp('foo', raw)